+++ src/host/alsa/mod.rs
@@ -634,8 +634,9 @@ pub struct Stream {
     trigger: TriggerSender,
 }
 
-// Compile-time assertion that Stream is Send
+// Compile-time assertion that Stream is Send and Sync
 crate::assert_stream_send!(Stream);
+crate::assert_stream_sync!(Stream);
 
 struct StreamWorkerContext {
     descriptors: Box<[libc::pollfd]>,