+++ src/host/jack/stream.rs
@@ -21,8 +21,9 @@ pub struct Stream {
     output_port_names: Vec<String>,
 }
 
-// Compile-time assertion that Stream is Send
+// Compile-time assertion that Stream is Send and Sync
 crate::assert_stream_send!(Stream);
+crate::assert_stream_sync!(Stream);
 
 impl Stream {
     // TODO: Return error messages